fix: extract Indicator onClick logic from UI level to custom hook and remove unused useLocalStorage
