+++ __tests__/components/carousel/Indicator.test.tsx
@@ -4,15 +4,7 @@ import { fireEvent, render, screen } from "@testing-library/react";
 describe("Carousel Indicator", () => {
   const setActiveIndex = jest.fn();
   it("should render indicator", () => {
-    render(
-      <Indicator
-        activeIndex={1}
-        length={3}
-        setActiveIndex={setActiveIndex}
-        setPreIndex={() => {}}
-        swiperInstance={null}
-      />
-    );
+    render(<Indicator activeIndex={1} length={3} handleIndicatorClick={() => {}} />);
 
     const indicators = screen.getAllByRole("indicator");
     expect(indicators).toBeInTheDocument;
@@ -20,20 +12,10 @@ describe("Carousel Indicator", () => {
   });
 
   it("should set correct active index when clicked", () => {
-    render(
-      <Indicator
-        activeIndex={1}
-        length={3}
-        setActiveIndex={setActiveIndex}
-        setPreIndex={() => {}}
-        swiperInstance={null}
-      />
-    );
+    render(<Indicator activeIndex={1} length={3} handleIndicatorClick={() => {}} />);
 
     const indicators = screen.getAllByRole("indicator");
     fireEvent.click(indicators[2]);
-
-    expect(setActiveIndex).toBeCalled();
     expect(indicators).toBeInTheDocument;
   });
 });
